+++ server/internal/adapter/integration/types.cfg.yaml
@@ -1,4 +1,4 @@
 package: integration
 generate:
   models: true
-output: ../../../pkg/integration/api/types.gen.go
+output: ../../../pkg/integrationapi/types.gen.go
